modify auto_delete_file_on_change

+++ products/models.py
@@ -91,10 +91,11 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
     except Product.DoesNotExist:
         return False
 
-    new_file = instance.avatar
-    if not old_file == new_file:
-        if os.path.exists(old_file.path):
-            os.remove(old_file.path)
+    if instance.avatar:
+        new_file = instance.avatar
+        if old_file and not old_file == new_file:
+            if os.path.exists(old_file.path):
+                os.remove(old_file.path)
 
 
 
@@ -113,10 +114,11 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
     except Category.DoesNotExist:
         return False
 
-    new_file = instance.avatar
-    if not old_file == new_file:
-        if os.path.exists(old_file.path):
-            os.remove(old_file.path)
+    if instance.avatar:
+        new_file = instance.avatar
+        if old_file and not old_file == new_file:
+            if os.path.exists(old_file.path):
+                os.remove(old_file.path)
 
 @receiver(models.signals.post_delete, sender=Category)
 def auto_delete_file_on_delete(sender, instance, **kwargs):
